Run through *all* resources in ProcessAsteroidDrill
In the common case (default Kerbalism config), Community Resource Pack
will attach multiple resources to asteroids.

KSP drills are kind of funky when there are multiple resources in the
game. When run as surface harvesters, they use ModuleResourceHarvester
and are specific to a single resource. However, when run as asteroid
harvesters, they will harvest all resources on the asteroid; there is no
resourceName property on ModuleAsteroidDrill.

The existing Kerbalism code was just shrugging and harvesting the first
ModuleAsteroidResource it found; dunno which one that would be.

The numbers are probably still all wrong.